move chunk graph building into separate file refactor chunk graph generation to include available modules optimization
