+++ lib/Chunk.js
@@ -355,7 +355,11 @@ class Chunk {
 		return this._modules.getFromUnorderedCache(getModulesIdent);
 	}
 
-	remove() {
+	/**
+	 * @param {string=} reason reason why chunk is removed
+	 * @returns {void}
+	 */
+	remove(reason) {
 		// cleanup modules
 		// Array.from is used here to create a clone, because removeChunk modifies this._modules
 		for (const module of Array.from(this._modules)) {
