+++ lib/ChunkGroup.js
@@ -345,7 +345,7 @@ class ChunkGroup {
 	}
 
 	/**
-	 * @param {ModuleReason} reason reason for removing ChunkGroup
+	 * @param {string=} reason reason for removing ChunkGroup
 	 * @returns {void}
 	 */
 	remove(reason) {
